refactor(server): object label identifier valdiation
