+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration/workspace-migration-builder/validators/services/flat-object-metadata-validator.service.ts
@@ -5,7 +5,6 @@ import { ALL_METADATA_NAME } from 'twenty-shared/metadata';
 import { isDefined } from 'twenty-shared/utils';
 
 import { findFlatEntityByUniversalIdentifier } from 'src/engine/metadata-modules/flat-entity/utils/find-flat-entity-by-universal-identifier.util';
-import { validateFlatObjectMetadataIdentifiers } from 'src/engine/metadata-modules/flat-object-metadata/validators/utils/validate-flat-object-metadata-identifiers.util';
 import { validateFlatObjectMetadataNameAndLabels } from 'src/engine/metadata-modules/flat-object-metadata/validators/utils/validate-flat-object-metadata-name-and-labels.util';
 import { ObjectMetadataExceptionCode } from 'src/engine/metadata-modules/object-metadata/object-metadata.exception';
 import { belongsToTwentyStandardApp } from 'src/engine/metadata-modules/utils/belongs-to-twenty-standard-app.util';
@@ -93,13 +92,6 @@ export class FlatObjectMetadataValidatorService {
           userFriendlyMessage: msg`Field label identifier is required`,
         });
       }
-
-      validationResult.errors.push(
-        ...validateFlatObjectMetadataIdentifiers({
-          universalFlatObjectMetadata: updatedFlatObjectMetadata,
-          universalFlatFieldMetadataMaps: flatFieldMetadataMaps,
-        }),
-      );
     }
 
     return validationResult;
@@ -214,12 +206,6 @@ export class FlatObjectMetadataValidatorService {
       });
     }
 
-    objectValidationResult.errors.push(
-      ...validateFlatObjectMetadataIdentifiers({
-        universalFlatObjectMetadata: flatObjectMetadataToValidate,
-        universalFlatFieldMetadataMaps: flatFieldMetadataMaps,
-      }),
-    );
     objectValidationResult.errors.push(
       ...validateFlatObjectMetadataNameAndLabels({
         optimisticUniversalFlatObjectMetadataMaps,
